w3c-reports/12092014/mac/runned-tests/html/semantics/embedded-content/media-elements/track/track-element/cors/030-expected.txt

Blocked access to external URL http://www1.localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/set-cookie.html#d9853a14-a7a3-4ae3-b94b-a0f187dfb10e
Blocked access to external URL http://www1.localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/cors-tester.py?id=d9853a14-a7a3-4ae3-b94b-a0f187dfb10e&origin=http%3A%2F%2Flocalhost%3A8800
CONSOLE MESSAGE: Cross-origin text track load denied by Cross-Origin Resource Sharing policy.
Blocked access to external URL http://www1.localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/cors-tester.py?delete-cookie&id=d9853a14-a7a3-4ae3-b94b-a0f187dfb10e
Actual	Expected
{
  "event": "error",
  "requests": [
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    }
  ]
}
{
  "event": "load",
  "requests": [
    {
      "cors": "no",
      "cookie": "yes"
    },
    {
      "cors": "http://localhost:8800",
      "cookie": "yes"
    }
  ]
}
Test id: d9853a14-a7a3-4ae3-b94b-a0f187dfb10e



FAIL track CORS: Use Credentials, same-origin, with headers, redirects to not same-origin, with headers assert_object_equals: property "event" expected object "[object Object]" got object "[object Object]"(stack: assert@http://localhost:8800/resources/testharness.js:1738:42
check_equal@http://localhost:8800/resources/testharness.js:378:28
assert_object_equals@http://localhost:8800/resources/testharness.js:390:21
checkData@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:128:29
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40
appendChild@[native code]
removeCookies@http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:104:34
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/html/semantics/embedded-content/media-elements/track/track-element/cors/support/common.js?pipe=sub:87:22
step@http://localhost:8800/resources/testharness.js:798:30
http://localhost:8800/resources/testharness.js:827:40)
